Implement backfill migration for tooltip format in properties and general sections
