--- server/data-migrations/1780385870217-BackfillTooltipFormat.ts
@@ -1,127 +1,0 @@
-import { MigrationInterface, QueryRunner } from 'typeorm';
-import { deleteAppHistoryForStructuralMigration } from '@helpers/migration.helper';
-
-const MIGRATION_NAME = 'BackfillTooltipFormat1780385870217';
-const BATCH_SIZE = 2000;
-
-const COMPONENT_TYPES = [
-  'Accordion',
-  'AudioRecorder',
-  'Button',
-  'ButtonGroupV2',
-  'Camera',
-  'Checkbox',
-  'CircularProgressBar',
-  'ColorPicker',
-  'CurrencyInput',
-  'DatePickerV2',
-  'DaterangePicker',
-  'DatetimePickerV2',
-  'Divider',
-  'DropdownV2',
-  'EmailInput',
-  'FileButton',
-  'FileInput',
-  'FilePicker',
-  'Form',
-  'Icon',
-  'IFrame',
-  'Image',
-  'JSONEditor',
-  'JSONExplorer',
-  'Kanban',
-  'KeyValuePair',
-  'Link',
-  'Listview',
-  'MultiselectV2',
-  'NumberInput',
-  'PasswordInput',
-  'PhoneInput',
-  'PopoverMenu',
-  'ProgressBar',
-  'RadioButtonV2',
-  'RangeSliderV2',
-  'ReorderableList',
-  'StarRating',
-  'Statistics',
-  'Tabs',
-  'Tags',
-  'TagsInput',
-  'Text',
-  'TextArea',
-  'TextInput',
-  'TimePicker',
-  'ToggleSwitchV2',
-  'TreeSelect',
-  'VerticalDivider',
-];
-
-const TOOLTIP_FORMAT_PATCH = JSON.stringify({
-  tooltipFormat: { value: 'plainText' },
-});
-
-export class BackfillTooltipFormat1780385870217 implements MigrationInterface {
-  public async up(queryRunner: QueryRunner): Promise<void> {
-    const [{ count }] = await queryRunner.query(
-      `SELECT COUNT(*) FROM components
-       WHERE type = ANY($1)
-         AND (properties IS NULL OR NOT (properties::jsonb ? 'tooltipFormat'))`,
-      [COMPONENT_TYPES]
-    );
-    const total = parseInt(count, 10);
-    console.log(`${MIGRATION_NAME}: [START] Backfill tooltipFormat | Total: ${total}`);
-
-    let lastId = '00000000-0000-0000-0000-000000000000';
-    let totalUpdated = 0;
-
-    while (true) {
-      const rows: { id: string }[] = await queryRunner.query(
-        `SELECT id FROM components
-         WHERE type = ANY($1)
-           AND id > $2
-           AND (properties IS NULL OR NOT (properties::jsonb ? $3))
-         ORDER BY id ASC
-         LIMIT $4`,
-        [COMPONENT_TYPES, lastId, 'tooltipFormat', BATCH_SIZE]
-      );
-
-      if (rows.length === 0) break;
-
-      lastId = rows[rows.length - 1].id;
-      const ids = rows.map((r) => r.id);
-
-      await queryRunner.query(
-        `UPDATE components
-         SET properties = (COALESCE(properties, '{}')::jsonb || $1::jsonb)::json
-         WHERE id = ANY($2::uuid[])`,
-        [TOOLTIP_FORMAT_PATCH, ids]
-      );
-
-      totalUpdated += rows.length;
-      const percentage = total > 0 ? ((totalUpdated / total) * 100).toFixed(1) : '0.0';
-      console.log(`${MIGRATION_NAME}: [PROGRESS] ${totalUpdated}/${total} (${percentage}%)`);
-    }
-
-    console.log(`${MIGRATION_NAME}: [SUCCESS] Backfill tooltipFormat finished.`);
-
-    if (totalUpdated > 0) {
-      const appVersionRows: { app_version_id: string }[] = await queryRunner.query(
-        `SELECT DISTINCT p.app_version_id
-         FROM components c
-         INNER JOIN pages p ON c.page_id = p.id
-         WHERE c.type = ANY($1)`,
-        [COMPONENT_TYPES]
-      );
-
-      await deleteAppHistoryForStructuralMigration(
-        queryRunner.manager,
-        { appVersionIds: appVersionRows.map((r) => r.app_version_id) },
-        MIGRATION_NAME
-      );
-    }
-  }
-
-  public async down(_queryRunner: QueryRunner): Promise<void> {
-    // Intentional no-op — structural migrations are not reversed
-  }
-}
